requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-asgi-connector==1.1.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bentoml==1.4.30
cattrs==23.1.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-option-group==0.5.9
cloudpickle==3.1.2
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.2
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
kantoku==0.18.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.5
nvidia-ml-py==13.580.82
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-aiohttp-client==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pip-requirements-parser==32.0.1
prometheus_client==0.23.1
propcache==0.4.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
schema==0.7.8
scikit-learn==1.7.2
scipy==1.16.3
simple-di==0.1.5
six==1.17.0
starlette==0.50.0
threadpoolctl==3.6.0
tomli_w==1.2.0
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-asgi-connector==1.1.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bentoml==1.4.30
cattrs==23.1.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-option-group==0.5.9
cloudpickle==3.1.2
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.2
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
kantoku==0.18.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.5
nvidia-ml-py==13.580.82
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-aiohttp-client==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pip-requirements-parser==32.0.1
prometheus_client==0.23.1
propcache==0.4.1
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
schema==0.7.8
scikit-learn==1.7.2
scipy==1.16.3
simple-di==0.1.5
six==1.17.0
starlette==0.50.0
threadpoolctl==3.6.0
tomli_w==1.2.0
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wrapt==1.17.3
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
